Split out unit tests, add further edge cases
